Renamed tests, Addded new AFE Get Test
